move unpinned version check to an observer
