+++ plugins/nf-schema/src/main/nextflow/validation/ValidationObserverFactory.groovy
@@ -10,7 +10,7 @@ class ValidationObserverFactory implements TraceObserverFactory {
 
     @Override
     Collection<TraceObserver> create(Session session) {
-        def List observers = []
+        def List observers = [ new ValidationObserver() ]
         // Only enable the help observer when a help message needs to be printed
         if(session.config.navigate('validation.help.enabled')) {
             observers.add(new HelpObserver())
